Create Overview
This is an overview for our game, like a mini design document

Game design documents include any information about art, narrative, design, and technical aspects that are needed to make the game happen.
Design docs are mostly used for reference to make sure the project is on track and are also act as a living document that can be revised at any time as the project evolves. 

I'm trying to keep this doc pretty lean since this is a jam game and we want to keep it as tight as possible.